Massive amount of changes. The first three options are now fully functional: fillPrescription();, viewStock();, and updateStock(); These three are as far as I understand fully functional.
Some changes that still need to be made:
If the user backspaces too far when entering password, the string array goes out of bounds and ends the program (I don't think I even remotely possess the knowledge to be able to fix this).
Bits of trace-data used to fix bugs where they have occurred throughout the program need to be found and taken out.
More console-clearings need to be established to create a cleaner interface for the user

MOST IMPORTANTLY:
Because of the way I have designed the receipt output (of which I find to be the most useful, having an individual file for each and every prescription filled where the name of the file is a combination of the patients name and the date the prescription was filled. Inside one can find all relevant information. It is not practical to write the code that would potentially open every single file and display it.